+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Added
 
 - `trimZeroValuedUnits` option in stringFromMilliseconds
+- `mathClamp` in common utils
 
 ## [2.9.0]