+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [2.10.1]
+
+### Fixed
+
+- `EnvVars`: error handling
+
 ## [2.10.0]
 
 ### Added